third_party/blink/web_tests/storage/websql/statement-success-callback-isolated-world-expected.txt

ALERT: undefined
ALERT: PASS: document.body.bar visible in a callback created in this world.
This test tests that the statement success callback is called in the right world.
